+++ pkg/kv/kvserver/kvstorage/BUILD.bazel
@@ -1,15 +1,17 @@
 load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "kvstorage",
     srcs = [
+        "cluster_version.go",
         "doc.go",
         "init.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvstorage",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/clusterversion",
         "//pkg/keys",
         "//pkg/kv/kvserver/logstore",
         "//pkg/roachpb",
@@ -20,6 +22,21 @@ go_library(
         "//pkg/util/log",
         "//pkg/util/protoutil",
         "@com_github_cockroachdb_errors//:errors",
+    ],
+)
+
+go_test(
+    name = "kvstorage_test",
+    srcs = ["cluster_version_test.go"],
+    args = ["-test.timeout=295s"],
+    embed = [":kvstorage"],
+    deps = [
+        "//pkg/clusterversion",
+        "//pkg/roachpb",
+        "//pkg/storage",
+        "//pkg/testutils",
+        "//pkg/util/leaktest",
+        "//pkg/util/log",
     ],
 )
