+++ pkg/server/migration.go
@@ -16,6 +16,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/clusterversion"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvstorage"
 	"github.com/cockroachdb/cockroach/pkg/server/serverpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
@@ -110,7 +111,7 @@ func bumpClusterVersion(
 ) error {
 
 	versionSetting := st.Version
-	prevCV, err := kvserver.SynthesizeClusterVersionFromEngines(
+	prevCV, err := kvstorage.SynthesizeClusterVersionFromEngines(
 		ctx, engines, versionSetting.BinaryVersion(),
 		versionSetting.BinaryMinSupportedVersion(),
 	)
@@ -130,7 +131,7 @@ func bumpClusterVersion(
 	// Whenever the version changes, we want to persist that update to
 	// wherever the CRDB process retrieved the initial version from
 	// (typically a collection of storage.Engines).
-	if err := kvserver.WriteClusterVersionToEngines(ctx, engines, newCV); err != nil {
+	if err := kvstorage.WriteClusterVersionToEngines(ctx, engines, newCV); err != nil {
 		return err
 	}
